fix and tests move to parser_test
